Refactor the Call trait into the test lib
